feat(worker): serve full MCP on Cloudflare Workers via Web Standard transport
Add a real Cloudflare Workers entrypoint that serves the complete IT Glue MCP
server over the SDK's Web Standard Streamable HTTP transport (runs natively on
workerd). IT Glue is called directly over global fetch (no vendor SDK), so the
server runs natively on the Workers runtime.

- Extract the side-effect-free server factory + ITGlueClient + helpers into
  src/mcp-server.ts; index.ts now owns only the Node-only transports (stdio +
  Node HTTP) and re-exports the shared API so existing consumers/tests keep
  working.
- src/worker.ts: stateless per-request server+transport, CORS, /health, /mcp,
  gateway-header and env credential resolution (X-ITGlue-API-Key / X-ITGlue-JWT
  / X-ITGlue-Region / X-ITGlue-Base-URL).
- wrangler.json: point main at src/worker.ts (wrangler bundles TS from source).
- src/__tests__/worker.test.ts (health, CORS, 404, initialize, tools/list,
  graceful no-creds error, gateway 401); gitignore .wrangler/.

Verified on real workerd (wrangler dev): /health ok, initialize returns
itglue-mcp, tools/list returns 20 tools. All 146 tests pass.
